+++ src/VBox/Runtime/r3/os2/mp-os2.cpp
@@ -1,4 +1,4 @@
-/* $Id: mp-os2.cpp 9429 2008-06-05 15:22:37Z knut.osmundsen@oracle.com $ */
+/* $Id: mp-os2.cpp 13908 2008-11-06 11:53:47Z noreply@oracle.com $ */
 /** @file
  * IPRT - Multiprocessor, OS/2.
  */
@@ -45,7 +45,7 @@
 
 RTDECL(int) RTMpCpuIdToSetIndex(RTCPUID idCpu)
 {
-    return idCpu < RTCPUSET_MAX_CPUS ? idCpu : -1;
+    return idCpu < RTCPUSET_MAX_CPUS ? (int) idCpu : -1;
 }
